*: decouple operator handler and server (#7089)
ref tikv/pd#5839
