+++ server/api/hot_status.go
@@ -22,6 +22,7 @@ import (
 	"strconv"
 
 	"github.com/tikv/pd/pkg/core"
+	"github.com/tikv/pd/pkg/errs"
 	"github.com/tikv/pd/pkg/statistics/buckets"
 	"github.com/tikv/pd/pkg/statistics/utils"
 	"github.com/tikv/pd/pkg/storage"
@@ -117,7 +118,7 @@ func (h *hotStatusHandler) GetHotWriteRegions(w http.ResponseWriter, r *http.Req
 		}
 		store := rc.GetStore(id)
 		if store == nil {
-			h.rd.JSON(w, http.StatusNotFound, server.ErrStoreNotFound(id).Error())
+			h.rd.JSON(w, http.StatusNotFound, errs.ErrStoreNotFound.FastGenByArgs(id).Error())
 			return
 		}
 		ids = append(ids, id)
@@ -153,7 +154,7 @@ func (h *hotStatusHandler) GetHotReadRegions(w http.ResponseWriter, r *http.Requ
 		}
 		store := rc.GetStore(id)
 		if store == nil {
-			h.rd.JSON(w, http.StatusNotFound, server.ErrStoreNotFound(id).Error())
+			h.rd.JSON(w, http.StatusNotFound, errs.ErrStoreNotFound.FastGenByArgs(id).Error())
 			return
 		}
 		ids = append(ids, id)
